requirements.txt
discord.py
python-dotenv
yt-dlp
static-ffmpeg
